// src/core/models/idea.ts
export interface Idea {
    id: number;
    title: string;
    category: string;
    category_display: string;
    status: string;
    status_display: string;
    latitude: string;
    longitude: string;
    author_email: string;
    zone_name: string;
    created_at: string;
    vote_count: number;
    positive_votes: number;
    negative_votes: number;
  }
  